+++ items/ration.lua
@@ -10,14 +10,17 @@ item:set_loot_tags(Item.LOOT_TAG.category_healing)
 
 item:onPickup(function(actor, stack)
     -- Restore all used Rations
-    local item      = Item.find("aphelion-ration")
     local item_used = Item.find("aphelion-rationUsed")
     local normal    = actor:item_stack_count(item_used, Item.TYPE.real)
     local temp      = actor:item_stack_count(item_used, Item.TYPE.temporary)
-    actor:item_remove(item_used, normal, false)
-    actor:item_remove(item_used, temp, true)
-    GM.item_give_internal(actor, item, normal, false)   -- Check if this still works in MP when the time comes
-    GM.item_give_internal(actor, item, temp, true)
+    if normal > 0 then
+        actor:item_remove(item_used, normal, false)
+        actor:item_give(item, normal, false)
+    end
+    if temp > 0 then
+        actor:item_remove(item_used, temp, true)
+        actor:item_give(item, temp, true)
+    end
 end)
 
 item:onDamaged(function(actor, damager, stack)
@@ -27,13 +30,16 @@ item:onDamaged(function(actor, damager, stack)
         gm.sound_play_at(sound, 0.9, 1.0, actor.x, actor.y, 1.0)
 
         -- Remove stacks and give used stacks
-        local item      = Item.find("aphelion-ration")
         local item_used = Item.find("aphelion-rationUsed")
         local normal    = actor:item_stack_count(item, Item.TYPE.real)
         local temp      = actor:item_stack_count(item, Item.TYPE.temporary)
-        actor:item_remove(item, normal, false)
-        actor:item_remove(item, temp, true)
-        GM.item_give_internal(actor, item_used, normal, false)
-        GM.item_give_internal(actor, item_used, temp, true)
+        if normal > 0 then
+            actor:item_remove(item, normal, false)
+            actor:item_give(item_used, normal, false)
+        end
+        if temp > 0 then
+            actor:item_remove(item, temp, true)
+            actor:item_give(item_used, temp, true)
+        end
     end
 end)