+++ items/sixShooter.lua
@@ -35,7 +35,7 @@ item:onAttack(function(actor, damager, stack)
         end
 
         if stack > 1 then
-            damager.damage = damager.damage * (0.75 + (0.25 * stack))
+            damager.damage = damager.damage * (0.67 + (0.33 * stack))
         end
     end
 end)