+++ equipment/thermiteFlare.lua
@@ -31,8 +31,8 @@
 -- -- Object
 
 -- local sprite = gm.constants.sBanditDynamite
--- -- local soundWindup = Resources.sfx_load(PATH.."assets/sounds/magicDaggerWindup.ogg")
--- -- local soundFreeze = Resources.sfx_load(PATH.."assets/sounds/magicDaggerFreeze.ogg")
+-- -- local soundWindup = Resources.sfx_load("aphelion", "", PATH.."assets/sounds/magicDaggerWindup.ogg")
+-- -- local soundFreeze = Resources.sfx_load("aphelion", "", PATH.."assets/sounds/magicDaggerFreeze.ogg")
 
 -- local obj = Object.new("aphelion", "thermiteFlareObject")
 -- obj:set_sprite(sprite)