+++ items/ration.lua
@@ -1,7 +1,7 @@
 -- Ration
 
 local sprite = Resources.sprite_load("aphelion", "ration", PATH.."assets/sprites/ration.png", 1, 16, 16)
-local sound = Resources.sfx_load(PATH.."assets/sounds/ration.ogg")
+local sound = Resources.sfx_load("aphelion", "ration", PATH.."assets/sounds/ration.ogg")
 
 local item = Item.new("aphelion", "ration")
 item:set_sprite(sprite)
@@ -10,14 +10,17 @@ item:set_loot_tags(Item.LOOT_TAG.category_healing)
 
 item:onPickup(function(actor, stack)
     -- Restore all used Rations
-    local item      = Item.find("aphelion-ration")
     local item_used = Item.find("aphelion-rationUsed")
     local normal    = actor:item_stack_count(item_used, Item.TYPE.real)
     local temp      = actor:item_stack_count(item_used, Item.TYPE.temporary)
-    actor:item_remove(item_used, normal, false)
-    actor:item_remove(item_used, temp, true)
-    gm.item_give_internal(actor.value, item.value, normal, false)   -- Check if this still works in MP when the time comes
-    gm.item_give_internal(actor.value, item.value, temp, true)
+    if normal > 0 then
+        actor:item_remove(item_used, normal, false)
+        actor:item_give(item, normal, false)
+    end
+    if temp > 0 then
+        actor:item_remove(item_used, temp, true)
+        actor:item_give(item, temp, true)
+    end
 end)
 
 item:onDamaged(function(actor, damager, stack)
@@ -27,13 +30,16 @@ item:onDamaged(function(actor, damager, stack)
         gm.sound_play_at(sound, 0.9, 1.0, actor.x, actor.y, 1.0)
 
         -- Remove stacks and give used stacks
-        local item      = Item.find("aphelion-ration")
         local item_used = Item.find("aphelion-rationUsed")
         local normal    = actor:item_stack_count(item, Item.TYPE.real)
         local temp      = actor:item_stack_count(item, Item.TYPE.temporary)
-        actor:item_remove(item, normal, false)
-        actor:item_remove(item, temp, true)
-        gm.item_give_internal(actor.value, item_used.value, normal, false)
-        gm.item_give_internal(actor.value, item_used.value, temp, true)
+        if normal > 0 then
+            actor:item_remove(item, normal, false)
+            actor:item_give(item_used, normal, false)
+        end
+        if temp > 0 then
+            actor:item_remove(item, temp, true)
+            actor:item_give(item_used, temp, true)
+        end
     end
 end)