+++ items/rationUsed.lua
@@ -6,24 +6,29 @@ local item = Item.new("aphelion", "rationUsed", true)
 item:set_sprite(sprite)
 
 item:onPickup(function(actor, stack)
-    actor.aphelion_ration_cooldown = 240 *60 * (1 - Helper.mixed_hyperbolic(stack, 0.2, 0))
+    actor:get_data("ration").cooldown = 240 *60 * (1 - Helper.mixed_hyperbolic(stack, 0.2, 0))
 end)
 
 local function restore_stacks(actor)
     -- Restore all used Rations
-    local item      = Item.find("aphelion-ration")
-    local item_used = Item.find("aphelion-rationUsed")
-    local normal    = actor:item_stack_count(item_used, Item.TYPE.real)
-    local temp      = actor:item_stack_count(item_used, Item.TYPE.temporary)
-    actor:item_remove(item_used, normal, false)
-    actor:item_remove(item_used, temp, true)
-    gm.item_give_internal(actor.value, item.value, normal, false)
-    gm.item_give_internal(actor.value, item.value, temp, true)
+    local item_ready = Item.find("aphelion-ration")
+    local normal    = actor:item_stack_count(item, Item.TYPE.real)
+    local temp      = actor:item_stack_count(item, Item.TYPE.temporary)
+    if normal > 0 then
+        actor:item_remove(item, normal, false)
+        actor:item_give(item_ready, normal, false)
+    end
+    if temp > 0 then
+        actor:item_remove(item, temp, true)
+        actor:item_give(item_ready, temp, true)
+    end
 end
 
 item:onStep(function(actor, stack)
+    local actorData = actor:get_data("ration")
+
     -- Tick down timer
-    if actor.aphelion_ration_cooldown > 0 then actor.aphelion_ration_cooldown = actor.aphelion_ration_cooldown - 1
+    if actorData.cooldown > 0 then actorData.cooldown = actorData.cooldown - 1
     else restore_stacks(actor)
     end
 end)