+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This update adds 2 skills.
 * Relic Guard
     * First stack: 50% -> 60%
     * Stacking: +25% -> +20%
+* Six Shooter : Stack 6th shot damage bonus: +25% -> +33%
+* Whimsical Star : Unlock requirement: 7 -> 9
 * Shattered Glass : No longer appears with Artifact of Enigma or in equipment activators.
 
 ### v1.1.4