+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This update adds 2 skills.
     * First stack: 50% -> 60%
     * Stacking: +25% -> +20%
 * Six Shooter : Stack 6th shot damage bonus: +25% -> +33%
+* Stiletto : Minor sprite update
 * Whimsical Star : Unlock requirement: 7 -> 9
 * Shattered Glass : No longer appears with Artifact of Enigma or in equipment activators.